requirements.txt
Flask==3.0.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
python-docx==0.8.11
